dummy_django_blog/requirements.txt
asgiref==3.8.1
astroid==3.2.3
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage>=6.5.0
coveralls>=3.3.1
defusedxml==0.7.1
dill==0.3.8
Django==4.2.14
django-bootstrap-icons==0.9.0
django-bootstrap-v5==1.0.11
flake8-report==0.1.0a1
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pylint==3.2.5
pytest==8.2.2
pytest-django==4.8.0
pytest-mock>=3.14.0
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.1
tomlkit==0.13.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
